+++ fabric/src/main/java/com/avp/fabric/data/compatibility/common/CommonConstants.java
@@ -49,6 +49,8 @@ public class CommonConstants {
 
     public static final TagKey<Item> RAW_MATERIALS_ALUMINUM = createItem("raw_materials/aluminum");
 
+    public static final TagKey<Item> RAW_MATERIALS_LEAD = createItem("raw_materials/lead");
+
     public static final TagKey<Item> URANIUM = createItem("ingots/uranium");
 
     /*